- Added "entity" to the search results, containing all the indexed fields, to make it easier to add field, - Added "class" to the search results entity, to allow Drupal to know what type of entity (item) it is, - Refactor the dummy search to be inline with the real search, - Added a "count" method to the ESClient interface, - Changed hits from int to long in the summary class, because the ElasticSearch "count" API returns long, - Refactor the summary to use "count" API instead of "search" API, for efficiency and get around the 10'000 limit of the search API, - Added index options to prevent the search from crashing when an index is missing, - Removed "start" from summary, I'm not sure why I added it in the first place.

+++ src/main/java/au/gov/aims/eatlas/searchengine/search/IndexSummary.java
@@ -23,7 +23,7 @@ import org.json.JSONObject;
 public class IndexSummary {
     private String index;
 
-    private Integer hits;
+    private Long hits;
 
     public String getIndex() {
         return this.index;
@@ -34,11 +34,11 @@ public class IndexSummary {
         return this;
     }
 
-    public Integer getHits() {
+    public Long getHits() {
         return this.hits;
     }
 
-    public IndexSummary setHits(Integer hits) {
+    public IndexSummary setHits(Long hits) {
         this.hits = hits;
         return this;
     }
